chore: Renamed Base Repository and Implementation of EntityTransformableRepository
