+++ src/domain/repositories/repository.interface.ts
@@ -3,7 +3,7 @@
  *
  * @template EntityType - The type of the entity.
  */
-export interface BaseRepository<EntityType> {
+export interface Repository<EntityType> {
   /**
    * Finds an entity by its ID.
    *
@@ -19,4 +19,6 @@ export interface BaseRepository<EntityType> {
    * @returns {Promise<void>} - A promise that resolves when the entity is saved.
    */
   save(entity: EntityType): Promise<void>
+
+
 }
